- Separate BagFile interface and implementation - Update dependencies - Remove dead code
